Cleaning up implementation CH-40999
