+++ Helper/Url.php
@@ -3,7 +3,6 @@
 namespace NS8\Protect\Helper;
 
 use Magento\Framework\App\Helper\AbstractHelper;
-use Magento\Framework\App\Request\Http;
 use Magento\Framework\UrlInterface;
 use Magento\Backend\Model\UrlInterface as BackendUrlInterface;
 use Magento\Framework\App\State;
@@ -39,34 +38,24 @@ class Url extends AbstractHelper
     protected $config;
 
     /**
-     * The request.
-     *
-     * @var Http
-     */
-    protected $request;
-
-    /**
      * The constructor.
      *
-     * @param BackendUrlInterface $backendUrl
-     * @param Config $config
-     * @param Http $request The request
-     * @param State $state
      * @param UrlInterface $url
+     * @param BackendUrlInterface $backendUrl
+     * @param State $state,
+     * @param Config $config
      *
      */
     public function __construct(
+        UrlInterface $url,
         BackendUrlInterface $backendUrl,
-        Config $config,
-        Http $request,
         State $state,
-        UrlInterface $url
+        Config $config
     ) {
-        $this->backendUrl = $backendUrl;
-        $this->config = $config;
-        $this->request = $request;
-        $this->state = $state;
         $this->url = $url;
+        $this->backendUrl = $backendUrl;
+        $this->state = $state;
+        $this->config = $config;
     }
 
     /**
@@ -129,40 +118,8 @@ class Url extends AbstractHelper
         return ClientSdk::getClientSdkUrl();
     }
 
-    /**
-     * Gets the protect-sdk-client URL
-     *
-     * @return string The URL
-     */
     public function getClientUrl(): string
     {
         return SdkConfigManager::getEnvValue('urls.client_url');
-    }
-
-    /**
-     * Get the page to navigate to within the protect client
-     *
-     * @return string The name of the page to naviage to.
-     */
-    public function getPageFromRequest(): string
-    {
-        $page = (string)$this->request->getParam('page');
-        $orderIncrementId = $this->getOrderIncrementIdFromRequest();
-        if (empty($page) && !empty($orderIncrementId)) {
-            $page = ClientSdkClient::CLIENT_PAGE_ORDER_DETAILS;
-        }
-
-        return $page;
-    }
-
-    /**
-     * Get the URL of the iframe that holds the NS8 Protect client.
-     *
-     * @return string The URL
-     */
-    public function getOrderIncrementIdFromRequest(): string
-    {
-        $orderId = $this->request->getParam('order_id');
-        return $orderId ? $this->order->getOrderIncrementId($orderId) : '';
     }
 }
